move heuristics into models.heuristics
